+++ paraflow-loader-producer/src/main/java/cn/edu/ruc/iir/paraflow/loader/producer/Producer.java
@@ -52,7 +52,7 @@ public interface Producer
                                                 List<String> columnName,
                                                 List<String> dataType);
 
-    void registerFiberFunc(String database, String table, Function<String, Long> fiberFunc);
+    void registerFiberFunc(String database, String table, Function<String, Integer> fiberFunc);
 
     void registerFilter(String database, String table, Function<Message, Boolean> filterFunc);
